NTBS-565: Move legacyView to LegacyNotifications page. Refactor
